requirements.txt
tensorboardX
future
lmdb
numpy
timm==0.5.4
opencv-python
pyyaml
requests
tqdm
yapf
pillow
scipy
scikit-image
imageio